Change to use S.Net.Socket and SslStream instead of StreamSocket on UAP
