+++ src/System.Private.ServiceModel/src/System/ServiceModel/Channels/IStreamUpgradeChannelBindingProvider.cs
@@ -2,19 +2,13 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-
-#if FEATURE_CORECLR // ExtendedProtection
 using System.Security.Authentication.ExtendedProtection;
-#endif
 
 namespace System.ServiceModel.Channels
 {
     internal interface IStreamUpgradeChannelBindingProvider : IChannelBindingProvider
     {
-#if FEATURE_CORECLR // ExtendedProtection
-
         ChannelBinding GetChannelBinding(StreamUpgradeInitiator upgradeInitiator, ChannelBindingKind kind);
-#endif
     }
 }
