+++ src/System.Private.ServiceModel/tests/Common/Unit/FakeAddress.cs
@@ -8,6 +8,8 @@ using System.Net;
 public static class FakeAddress
 {
     public static string HttpAddress = "http://" + IPAddress.None + ":0";
+    
+    public static string HttpsAddress = "https://" + IPAddress.None + ":0";
 
     public static string TcpAddress = "net.tcp://" + IPAddress.None + ":0";
 }
